+++ sdk/keyvault/keyvault-certificates/src/index.ts
@@ -139,7 +139,6 @@ import {
 } from "./core/models";
 import { KeyVaultClient } from "./core/keyVaultClient";
 import { SDK_VERSION } from "./core/utils/constants";
-import { parseKeyvaultIdentifier as parseKeyvaultEntityIdentifier } from "./core/utils";
 import "@azure/core-paging";
 import { PageSettings, PagedAsyncIterableIterator } from "@azure/core-paging";
 import { challengeBasedAuthenticationPolicy } from "../../keyvault-common/src";
@@ -152,6 +151,7 @@ import { DeleteCertificateState } from "./lro/delete/operation";
 import { CreateCertificateState } from "./lro/create/operation";
 import { RecoverDeletedCertificateState } from "./lro/recover/operation";
 import { parseCertificateBytes } from "./utils";
+import { KeyVaultCertificatesIdentifier, ParsedKeyVaultCertificatesIdentifier } from "./identifier";
 
 export {
   CertificateClientOptions,
@@ -164,6 +164,7 @@ export {
   BeginRecoverDeletedCertificateOptions,
   KeyVaultCertificate,
   KeyVaultCertificateWithPolicy,
+  KeyVaultCertificatesIdentifier,
   BackupCertificateOptions,
   CertificateContentType,
   CertificateProperties,
@@ -177,6 +178,7 @@ export {
   CertificateTags,
   CreateCertificateOptions,
   CertificatePollerOptions,
+  ParsedKeyVaultCertificatesIdentifier,
   PollerLike,
   CreateCertificateState,
   DeleteCertificateState,
@@ -369,7 +371,7 @@ function toPublicPolicy(policy: CoreCertificatePolicy = {}): CertificatePolicy {
 }
 
 function toPublicIssuer(issuer: IssuerBundle = {}): CertificateIssuer {
-  const parsedId = parseKeyvaultEntityIdentifier("certificates", issuer.id);
+  const parsedId = new KeyVaultCertificatesIdentifier(issuer.id!);
   const attributes: IssuerAttributes = issuer.attributes || {};
 
   const publicIssuer: CertificateIssuer = {
@@ -490,7 +492,7 @@ export class CertificateClient {
       const currentSetResponse = await this.client.getCertificates(this.vaultUrl, optionsComplete);
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle);
+        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle, this);
       }
     }
     while (continuationState.continuationToken) {
@@ -500,7 +502,7 @@ export class CertificateClient {
       );
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle);
+        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle, this);
       } else {
         break;
       }
@@ -582,7 +584,7 @@ export class CertificateClient {
       );
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle);
+        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle, this);
       }
     }
     while (continuationState.continuationToken) {
@@ -593,7 +595,7 @@ export class CertificateClient {
       );
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle);
+        yield currentSetResponse.value.map(this.getPropertiesFromCertificateBundle, this);
       } else {
         break;
       }
@@ -1438,15 +1440,10 @@ export class CertificateClient {
     let result: UpdateCertificateResponse;
 
     try {
-      result = await this.client.updateCertificate(
-        this.vaultUrl,
-        certificateName,
-        version,
-        {
-          ...this.setParentSpan(span, requestOptions),
-          certificateAttributes: toCoreAttributes(options)
-        }
-      );
+      result = await this.client.updateCertificate(this.vaultUrl, certificateName, version, {
+        ...this.setParentSpan(span, requestOptions),
+        certificateAttributes: toCoreAttributes(options)
+      });
     } finally {
       span.end();
     }
@@ -1724,7 +1721,7 @@ export class CertificateClient {
       );
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getDeletedCertificateFromItem);
+        yield currentSetResponse.value.map(this.getDeletedCertificateFromItem, this);
       }
     }
     while (continuationState.continuationToken) {
@@ -1734,7 +1731,7 @@ export class CertificateClient {
       );
       continuationState.continuationToken = currentSetResponse.nextLink;
       if (currentSetResponse.value) {
-        yield currentSetResponse.value.map(this.getDeletedCertificateFromItem);
+        yield currentSetResponse.value.map(this.getDeletedCertificateFromItem, this);
       } else {
         break;
       }
@@ -2015,7 +2012,7 @@ export class CertificateClient {
   private getPropertiesFromCertificateBundle(
     certificateBundle: CertificateBundle
   ): CertificateProperties {
-    const parsedId = parseKeyvaultEntityIdentifier("certificates", certificateBundle.id);
+    const parsedId = new KeyVaultCertificatesIdentifier(certificateBundle.id!);
     const attributes: CertificateAttributes = certificateBundle.attributes || {};
 
     const abstractProperties: CertificateProperties = {
@@ -2073,7 +2070,7 @@ export class CertificateClient {
   private getCertificateFromCertificateBundle(
     certificateBundle: CertificateBundle
   ): KeyVaultCertificate {
-    const parsedId = parseKeyvaultEntityIdentifier("certificates", certificateBundle.id);
+    const parsedId = new KeyVaultCertificatesIdentifier(certificateBundle.id!);
 
     const attributes: CertificateAttributes = certificateBundle.attributes || {};
 
@@ -2105,7 +2102,7 @@ export class CertificateClient {
   private getCertificateWithPolicyFromCertificateBundle(
     certificateBundle: CertificateBundle
   ): KeyVaultCertificateWithPolicy {
-    const parsedId = parseKeyvaultEntityIdentifier("certificates", certificateBundle.id);
+    const parsedId = new KeyVaultCertificatesIdentifier(certificateBundle.id!);
 
     const attributes: CertificateAttributes = certificateBundle.attributes || {};
     const policy = toPublicPolicy(certificateBundle.policy || {});
@@ -2152,7 +2149,7 @@ export class CertificateClient {
   }
 
   private getDeletedCertificateFromItem(item: DeletedCertificateItem): DeletedCertificate {
-    const parsedId = parseKeyvaultEntityIdentifier("certificates", item.id);
+    const parsedId = new KeyVaultCertificatesIdentifier(item.id!);
 
     const attributes: any = item.attributes || {};